+++ core/src/lib.rs
@@ -96,19 +96,6 @@
 
 #![cfg_attr(not(feature = "std"), no_std)]
 
-use core::cmp::min;
-use core::ops::{Deref, DerefMut};
-use core::ptr::NonNull;
-use core::result::Result as CoreResult;
-use core::slice::from_raw_parts;
-use core::slice::from_raw_parts_mut;
-use core::sync::atomic::{
-    AtomicBool, AtomicUsize,
-    Ordering::{Acquire, Relaxed, Release},
-};
-
-pub mod ng;
-
 /// Result type used by the `BBQueue` interfaces
 pub type Result<T> = CoreResult<T, Error>;
 
@@ -122,11 +109,63 @@ pub enum Error {
     Refactoring,
 }
 
-/// A single producer, single consumer, thread safe queue
-#[derive(Debug)]
-pub struct BBQueue {
+use core::ops::{Deref, DerefMut};
+use core::result::Result as CoreResult;
+use core::cell::UnsafeCell;
+use core::marker::PhantomData;
+use core::mem::{size_of, MaybeUninit};
+use core::ptr::NonNull;
+use core::slice::from_raw_parts;
+use core::slice::from_raw_parts_mut;
+use core::sync::atomic::{
+    AtomicBool, AtomicUsize,
+    Ordering::{Acquire, Relaxed, Release},
+};
+use generic_array::{ArrayLength, GenericArray};
+
+/// A backing structure for a BBQueue. Can be used to create either
+/// a BBQueue or a split Producer/Consumer pair
+pub struct BBBuffer<N: ArrayLength<u8>> {
     // Underlying data storage
-    buf: NonNull<[u8]>,
+    inner: ConstBBBuffer<GenericArray<u8, N>>,
+}
+
+impl<N> BBBuffer<N>
+where
+    N: ArrayLength<u8>,
+{
+    pub fn try_split(&self) -> Result<(Producer<N>, Consumer<N>)> {
+        if self.inner.already_split.swap(true, Relaxed) {
+            return Err(Error::Refactoring);
+        } else {
+            unsafe {
+                // Explicitly zero the data to avoid undefined behavior.
+                // This is required, because we hand out references to the buffers,
+                // which mean that creating them as references is technically UB for now
+                let mu_ptr = self.inner.buf.get();
+                (*mu_ptr).as_mut_ptr().write_bytes(0u8, 1);
+
+                let nn1 = NonNull::new_unchecked(self as *const _ as *mut _);
+                let nn2 = NonNull::new_unchecked(self as *const _ as *mut _);
+
+                Ok((
+                    Producer {
+                        bbq: nn1,
+                        pd: PhantomData,
+                    },
+                    Consumer {
+                        bbq: nn2,
+                        pd: PhantomData,
+                    },
+                ))
+            }
+        }
+    }
+}
+
+/// `const-fn` version BBBuffer
+pub struct ConstBBBuffer<A> {
+    buf: UnsafeCell<MaybeUninit<A>>,
 
     /// Where the next byte will be written
     write: AtomicUsize,
@@ -154,22 +193,12 @@ pub struct BBQueue {
     already_split: AtomicBool,
 }
 
-impl BBQueue {
-    /// Create a new BBQueue with a given backing buffer. After giving this
-    /// buffer to `BBQueue::new(), the backing buffer must not be used, or
-    /// undefined behavior could occur!
-    ///
-    /// Additionally, when using `BBQueue::split()`, the `BBQueue` struct must
-    /// never be moved, otherwise `Producer` and `Consumer` could corrupt memory
-    ///
-    /// Consider using the `bbq!()` macro to safely create a statically
-    /// allocated instance, or enable the "std" feature, and instead use the
-    /// `BBQueue::new_boxed()` constructor.
-    pub unsafe fn unpinned_new(buf: &'static mut [u8]) -> Self {
-        let sz = buf.len();
-        assert!(sz != 0);
-        BBQueue {
-            buf: NonNull::new_unchecked(buf),
+impl<A> ConstBBBuffer<A> {
+    pub const fn new() -> Self {
+        Self {
+            // Note, we contain u8's, so zeroing is a sound strategy for
+            // initialization
+            buf: UnsafeCell::new(MaybeUninit::uninit()),
 
             /// Owned by the writer
             write: AtomicUsize::new(0),
@@ -178,7 +207,7 @@ impl BBQueue {
             read: AtomicUsize::new(0),
 
             /// Cooperatively owned
-            last: AtomicUsize::new(sz),
+            last: AtomicUsize::new(size_of::<A>()),
 
             /// Owned by the Writer, "private"
             reserve: AtomicUsize::new(0),
@@ -189,24 +218,39 @@ impl BBQueue {
             already_split: AtomicBool::new(false),
         }
     }
+}
 
+pub struct Producer<'a, N>
+where
+    N: ArrayLength<u8>,
+{
+    bbq: NonNull<BBBuffer<N>>,
+    pd: PhantomData<&'a ()>,
+}
+
+impl<'a, N> Producer<'a, N>
+where
+    N: ArrayLength<u8>,
+{
     /// Request a writable, contiguous section of memory of exactly
     /// `sz` bytes. If the buffer size requested is not available,
     /// an error will be returned.
     pub fn grant(&mut self, sz: usize) -> Result<GrantW> {
+        let inner = unsafe { &self.bbq.as_ref().inner };
+
         // Writer component. Must never write to `read`,
         // be careful writing to `load`
+        let write = inner.write.load(Relaxed);
 
-        let write = self.write.load(Relaxed);
-
-        if self.reserve.load(Relaxed) != write {
+        if inner.reserve.load(Relaxed) != write {
             // GRANT IN PROCESS, do not allow further grants
             // until the current one has been completed
             return Err(Error::GrantInProgress);
         }
 
-        let read = self.read.load(Acquire);
-        let max = unsafe { self.buf.as_mut().len() };
+        let read = inner.read.load(Acquire);
+
+        let max = N::to_usize();
 
         let already_inverted = write < read;
 
@@ -239,86 +283,89 @@ impl BBQueue {
         };
 
         // Safe write, only viewed by this task
-        self.reserve.store(start + sz, Relaxed);
+        inner.reserve.store(start + sz, Relaxed);
 
-        let c = unsafe { self.buf.as_mut().as_mut_ptr() };
-        let d = unsafe { from_raw_parts_mut(c, max) };
+        let c = unsafe { (*inner.buf.get()).as_mut_ptr().cast::<u8>() };
+        let d = unsafe { from_raw_parts_mut(c.offset(start as isize), inner.reserve.load(Relaxed)) };
 
         Ok(GrantW {
-            buf: &mut d[start..self.reserve.load(Relaxed)],
+            buf: d,
         })
     }
 
-    /// Request a writable, contiguous section of memory of up to
-    /// `sz` bytes. If a buffer of size `sz` is not available, but
-    /// some space (0 < available < sz) is available, then a grant
-    /// will be given for the remaining size. If no space is available
-    /// for writing, an error will be returned
-    pub fn grant_max(&mut self, mut sz: usize) -> Result<GrantW> {
-        // Writer component. Must never write to `read`,
-        // be careful writing to `load`
+    //     /// Request a writable, contiguous section of memory of up to
+    //     /// `sz` bytes. If a buffer of size `sz` is not available, but
+    //     /// some space (0 < available < sz) is available, then a grant
+    //     /// will be given for the remaining size. If no space is available
+    //     /// for writing, an error will be returned
+    //     fn grant_max(&mut self, mut sz: usize) -> Result<GrantW> {
+    //         // Writer component. Must never write to `read`,
+    //         // be careful writing to `load`
 
-        let write = self.write.load(Relaxed);
+    //         let write = self.inner.write.load(Relaxed);
 
-        if self.reserve.load(Relaxed) != write {
-            // GRANT IN PROCESS, do not allow further grants
-            // until the current one has been completed
-            return Err(Error::GrantInProgress);
-        }
+    //         if self.inner.reserve.load(Relaxed) != write {
+    //             // GRANT IN PROCESS, do not allow further grants
+    //             // until the current one has been completed
+    //             return Err(Error::GrantInProgress);
+    //         }
 
-        let read = self.read.load(Acquire);
-        let max = unsafe { self.buf.as_mut().len() };
+    //         let read = self.inner.read.load(Acquire);
+    //         let max = unsafe { (*self.inner.buf.as_mut_ptr()).as_mut().len() };
 
-        let already_inverted = write < read;
+    //         let already_inverted = write < read;
 
-        let start = if already_inverted {
-            // In inverted case, read is always > write
-            let remain = read - write - 1;
+    //         let start = if already_inverted {
+    //             // In inverted case, read is always > write
+    //             let remain = read - write - 1;
 
-            if remain != 0 {
-                sz = min(remain, sz);
-                write
-            } else {
-                // Inverted, no room is available
-                return Err(Error::InsufficientSize);
-            }
-        } else {
-            if write != max {
-                // Some (or all) room remaining in un-inverted case
-                sz = min(max - write, sz);
-                write
-            } else {
-                // Not inverted, but need to go inverted
+    //             if remain != 0 {
+    //                 sz = min(remain, sz);
+    //                 write
+    //             } else {
+    //                 // Inverted, no room is available
+    //                 return Err(Error::InsufficientSize);
+    //             }
+    //         } else {
+    //             if write != max {
+    //                 // Some (or all) room remaining in un-inverted case
+    //                 sz = min(max - write, sz);
+    //                 write
+    //             } else {
+    //                 // Not inverted, but need to go inverted
 
-                // NOTE: We check read > 1, NOT read >= 1, because
-                // write must never == read in an inverted condition, since
-                // we will then not be able to tell if we are inverted or not
-                if read > 1 {
-                    sz = min(read - 1, sz);
-                    0
-                } else {
-                    // Not invertible, no space
-                    return Err(Error::InsufficientSize);
-                }
-            }
-        };
+    //                 // NOTE: We check read > 1, NOT read >= 1, because
+    //                 // write must never == read in an inverted condition, since
+    //                 // we will then not be able to tell if we are inverted or not
+    //                 if read > 1 {
+    //                     sz = min(read - 1, sz);
+    //                     0
+    //                 } else {
+    //                     // Not invertible, no space
+    //                     return Err(Error::InsufficientSize);
+    //                 }
+    //             }
+    //         };
 
-        // Safe write, only viewed by this task
-        self.reserve.store(start + sz, Relaxed);
+    //         // Safe write, only viewed by this task
+    //         self.inner.reserve.store(start + sz, Relaxed);
 
-        let c = unsafe { self.buf.as_mut().as_mut_ptr() };
-        let d = unsafe { from_raw_parts_mut(c, max) };
+    //         let c = unsafe { (*self.inner.buf.as_mut_ptr()).as_mut().as_mut_ptr() };
+    //         let d = unsafe { from_raw_parts_mut(c, max) };
 
-        Ok(GrantW {
-            buf: &mut d[start..self.reserve.load(Relaxed)],
-        })
-    }
+    //         Ok(GrantW {
+    //             buf: &mut d[start..self.inner.reserve.load(Relaxed)],
+    //         })
+    //     }
+
 
     /// Finalizes a writable grant given by `grant()` or `grant_max()`.
     /// This makes the data available to be read via `read()`.
     ///
     /// If `used` is larger than the given grant, this function will panic.
     pub fn commit(&mut self, used: usize, grant: GrantW) {
+        let inner = unsafe { &self.bbq.as_ref().inner };
+
         // Writer component. Must never write to READ,
         // be careful writing to LAST
 
@@ -328,41 +375,56 @@ impl BBQueue {
         assert!(len >= used);
 
         // Verify we are committing OUR grant
-        assert!(self.is_our_grant(&grant.buf));
+        assert!(unsafe { self.bbq.as_ref().is_our_grant(&grant.buf) });
 
         drop(grant);
 
-        let write = self.write.load(Relaxed);
-        self.reserve.fetch_sub(len - used, Relaxed);
+        let write = inner.write.load(Relaxed);
+        inner.reserve.fetch_sub(len - used, Relaxed);
 
-        let max = unsafe { self.buf.as_mut().len() };
-        let last = self.last.load(Relaxed);
+        let max = N::to_usize();
+        let last = inner.last.load(Relaxed);
 
         // Inversion case, we have begun writing
-        if (self.reserve.load(Relaxed) < write) && (write != max) {
-            self.last.store(write, Release);
+        if (inner.reserve.load(Relaxed) < write) && (write != max) {
+            inner.last.store(write, Release);
         } else if write > last {
-            self.last.store(max, Release);
+            inner.last.store(max, Release);
         }
 
         // Write must be updated AFTER last, otherwise read could think it was
         // time to invert early!
-        self.write.store(self.reserve.load(Relaxed), Release);
+        inner.write.store(inner.reserve.load(Relaxed), Release);
     }
+}
+
+pub struct Consumer<'a, N>
+where
+    N: ArrayLength<u8>,
+{
+    bbq: NonNull<BBBuffer<N>>,
+    pd: PhantomData<&'a ()>,
+}
+
+impl<'a, N> Consumer<'a, N>
+where
+    N: ArrayLength<u8>,
+{
 
     /// Obtains a contiguous slice of committed bytes. This slice may not
     /// contain ALL available bytes, if the writer has wrapped around. The
     /// remaining bytes will be available after all readable bytes are
     /// released
     pub fn read(&mut self) -> Result<GrantR> {
-        if self.read_in_progress.load(Relaxed) {
+        let inner = unsafe { &self.bbq.as_ref().inner };
+
+        if inner.read_in_progress.load(Relaxed) {
             return Err(Error::GrantInProgress);
         }
 
-        let write = self.write.load(Acquire);
-        let last = self.last.load(Acquire);
-        let mut read = self.read.load(Relaxed);
-        let max = unsafe { self.buf.as_mut().len() };
+        let write = inner.write.load(Acquire);
+        let last = inner.last.load(Acquire);
+        let mut read = inner.read.load(Relaxed);
 
         // Resolve the inverted case or end of read
         if (read == last) && (write < read) {
@@ -375,7 +437,7 @@ impl BBQueue {
             //   Commit does not check read, but if Grant has started an inversion,
             //   grant could move Last to the prior write position
             // MOVING READ BACKWARDS!
-            self.read.store(0, Release);
+            inner.read.store(0, Release);
         }
 
         let sz = if write < read {
@@ -390,13 +452,13 @@ impl BBQueue {
             return Err(Error::InsufficientSize);
         }
 
-        self.read_in_progress.store(true, Relaxed);
+        inner.read_in_progress.store(true, Relaxed);
 
-        let c = unsafe { self.buf.as_mut().as_ptr() };
-        let d = unsafe { from_raw_parts(c, max) };
+        let c = unsafe { (*inner.buf.get()).as_ptr().cast::<u8>() };
+        let d = unsafe { from_raw_parts(c.offset(read as isize), sz) };
 
         Ok(GrantR {
-            buf: &d[read..read + sz],
+            buf: d,
         })
     }
 
@@ -404,69 +466,76 @@ impl BBQueue {
     /// to be used by later writes
     ///
     /// If `used` is larger than the given grant, this function will panic.
-    pub fn release(&self, used: usize, grant: GrantR) {
+    pub fn release(&mut self, used: usize, grant: GrantR) {
+        let inner = unsafe { &self.bbq.as_ref().inner };
+
         assert!(used <= grant.buf.len());
 
         // Verify we are committing OUR grant
-        assert!(self.is_our_grant(&grant.buf));
+        assert!(unsafe { self.bbq.as_ref().is_our_grant(&grant.buf) });
 
         drop(grant);
 
         // This should be fine, purely incrementing
-        let _ = self.read.fetch_add(used, Release);
+        let _ = inner.read.fetch_add(used, Release);
 
-        self.read_in_progress.store(false, Relaxed);
+        inner.read_in_progress.store(false, Relaxed);
     }
+}
+
+// Private impls, used by Queue or Producer/Consumer
+impl<N> BBBuffer<N>
+where
+    N: ArrayLength<u8>,
+{
 
     /// Returns the size of the backing storage.
     ///
     /// This is the maximum number of bytes that can be stored in this queue.
     pub fn capacity(&self) -> usize {
-        unsafe { self.buf.as_ref().len() }
+        N::to_usize()
     }
 
-    fn is_our_grant(&self, gr_buf: &[u8]) -> bool {
-        let start = unsafe { self.buf.as_ref().as_ptr() as usize };
-        let end_plus_one = start + unsafe { self.buf.as_ref().len() };
-        let buf_start = gr_buf.as_ptr() as usize;
-        let buf_end_plus_one = buf_start + gr_buf.len();
 
-        (start <= buf_start) && (buf_end_plus_one <= end_plus_one)
-    }
-}
+    pub(crate) fn is_our_grant(&self, gr_buf: &[u8]) -> bool {
+        let buf_start = self.inner.buf.get() as usize;
+        let gr_start = gr_buf.as_ptr() as usize;
+        let buf_end_plus_one = buf_start + N::to_usize();
+        let gr_end_plus_one = gr_start + gr_buf.len();
 
-impl BBQueue {
-    /// This method takes a `BBQueue`, and returns a set of SPSC handles
-    /// that may be given to separate threads. May only be called once
-    /// per `BBQueue` object, or this function will panic
-    pub fn split(&self) -> (Producer, Consumer) {
-        assert!(!self.already_split.swap(true, Relaxed));
-
-        let nn1 = unsafe { NonNull::new_unchecked(self as *const _ as *mut _) };
-        let nn2 = unsafe { NonNull::new_unchecked(self as *const _ as *mut _) };
-
-        (Producer { bbq: nn1 }, Consumer { bbq: nn2 })
+        (buf_start <= gr_start) && (gr_end_plus_one <= buf_end_plus_one)
     }
 }
 
-#[cfg(feature = "std")]
-impl BBQueue {
-    /// Creates a Boxed `BBQueue`
-    ///
-    /// NOTE: This function essentially "leaks" the backing buffer
-    /// (e.g. `BBQueue::new_boxed(1024)` will leak 1024 bytes). This
-    /// may be changed in the future.
-    pub fn new_boxed(capacity: usize) -> Box<Self> {
-        let data: &mut [u8] = Box::leak(vec![0; capacity].into_boxed_slice());
-        Box::new(unsafe { Self::unpinned_new(data) })
-    }
-
-    /// Splits a boxed `BBQueue` into a producer and consumer
-    pub fn split_box(queue: Box<Self>) -> (Producer, Consumer) {
-        let self_ref = Box::leak(queue);
-        Self::split(self_ref)
+impl<N> BBBuffer<N>
+where
+    N: ArrayLength<u8>,
+{
+    pub fn new() -> Self {
+        Self {
+            inner: ConstBBBuffer::new(),
+        }
     }
 }
+
+// #[cfg(feature = "std")]
+// impl BBQueue {
+//     /// Creates a Boxed `BBQueue`
+//     ///
+//     /// NOTE: This function essentially "leaks" the backing buffer
+//     /// (e.g. `BBQueue::new_boxed(1024)` will leak 1024 bytes). This
+//     /// may be changed in the future.
+//     pub fn new_boxed(capacity: usize) -> Box<Self> {
+//         let data: &mut [u8] = Box::leak(vec![0; capacity].into_boxed_slice());
+//         Box::new(unsafe { Self::unpinned_new(data) })
+//     }
+
+//     /// Splits a boxed `BBQueue` into a producer and consumer
+//     pub fn split_box(queue: Box<Self>) -> (Producer, Consumer) {
+//         let self_ref = Box::leak(queue);
+//         Self::split(self_ref)
+//     }
+// }
 
 /// A structure representing a contiguous region of memory that
 /// may be written to, and potentially "committed" to the queue
@@ -517,147 +586,69 @@ impl Deref for GrantR {
     }
 }
 
-unsafe impl Send for Consumer {}
 
-/// An opaque structure, capable of reading data from the queue
-pub struct Consumer {
-    /// The underlying `BBQueue` object`
-    bbq: NonNull<BBQueue>,
-}
+// /// Statically allocate a `BBQueue` singleton object with a given size (in bytes).
+// ///
+// /// This function must only ever be called once in the same place
+// /// per function. For example:
+// ///
+// /// ```rust,ignore
+// /// // this creates two separate/distinct buffers, and is an acceptable usage
+// /// let bbq1 = bbq!(1024).unwrap(); // Returns Some(BBQueue)
+// /// let bbq2 = bbq!(1024).unwrap(); // Returns Some(BBQueue)
+// ///
+// /// // this is a bad example, as the same instance is executed twice!
+// /// // On the first call, this macro will return `Some(BBQueue)`. On the
+// /// // second call, this macro will return `None`!
+// /// for _ in 0..2 {
+// ///     let _ = bbq!(1024).unwrap();
+// /// }
+// /// ```
+// ///
+// /// If you are using this in a cortex-m microcontroller system,
+// /// consider using the cortex_m_bbq!() macro instead.
+// #[macro_export]
+// macro_rules! bbq {
+//     ($expr:expr) => {{
+//         use core::sync::atomic::{AtomicBool, Ordering};
 
-unsafe impl Send for Producer {}
+//         static TAKEN: AtomicBool = AtomicBool::new(false);
 
-/// An opaque structure, capable of writing data to the queue
-pub struct Producer {
-    /// The underlying `BBQueue` object`
-    bbq: NonNull<BBQueue>,
-}
+//         if TAKEN.compare_and_swap(false, true, Ordering::AcqRel) {
+//             None
+//         } else {
+//             unsafe { $crate::unchecked_bbq!($expr) }
+//         }
+//     }};
+// }
 
-impl Producer {
-    /// Request a writable, contiguous section of memory of exactly
-    /// `sz` bytes. If the buffer size requested is not available,
-    /// an error will be returned.
-    #[inline(always)]
-    pub fn grant(&mut self, sz: usize) -> Result<GrantW> {
-        unsafe { self.bbq.as_mut().grant(sz) }
-    }
+// /// Like the `bbq!()` macro, but wraps the initialization in a cortex-m
+// /// "critical section" by disabling interrupts
+// #[cfg_attr(feature = "cortex-m", macro_export)]
+// #[allow(unused_macros)]
+// macro_rules! cortex_m_bbq {
+//     ($expr:expr) => {
+//         cortex_m::interrupt::free(|_| unsafe { $crate::unchecked_bbq!($expr) })
+//     };
+// }
 
-    /// Request a writable, contiguous section of memory of up to
-    /// `sz` bytes. If a buffer of size `sz` is not available, but
-    /// some space (0 < available < sz) is available, then a grant
-    /// will be given for the remaining size. If no space is available
-    /// for writing, an error will be returned
-    #[inline(always)]
-    pub fn grant_max(&mut self, sz: usize) -> Result<GrantW> {
-        unsafe { self.bbq.as_mut().grant_max(sz) }
-    }
+// /// This macro does try to provide similar guarantees as `bbq!()`,
+// /// but is not thread safe.
+// #[macro_export]
+// macro_rules! unchecked_bbq {
+//     ($expr:expr) => {{
+//         static mut BUFFER: [u8; $expr] = [0u8; $expr];
 
-    /// Finalizes a writable grant given by `grant()` or `grant_max()`.
-    /// This makes the data available to be read via `read()`.
-    ///
-    /// If `used` is larger than the given grant, this function will panic.
-    #[inline(always)]
-    pub fn commit(&mut self, used: usize, grant: GrantW) {
-        unsafe { self.bbq.as_mut().commit(used, grant) }
-    }
+//         // Really, this shouldn't be an option. But for now, there is
+//         // no stable way to get an uninitialized static, so we pay the
+//         // option cost for compatibility
+//         static mut BBQ: Option<BBQueue> = None;
 
-    /// Returns the size of the backing storage.
-    ///
-    /// This is the maximum number of bytes that can be stored in this queue.
-    pub fn capacity(&self) -> usize {
-        unsafe { self.bbq.as_ref().capacity() }
-    }
-}
-
-impl Consumer {
-    /// Obtains a contiguous slice of committed bytes. This slice may not
-    /// contain ALL available bytes, if the writer has wrapped around. The
-    /// remaining bytes will be available after all readable bytes are
-    /// released
-    #[inline(always)]
-    pub fn read(&mut self) -> Result<GrantR> {
-        unsafe { self.bbq.as_mut().read() }
-    }
-
-    /// Release a sequence of bytes from the buffer, allowing the space
-    /// to be used by later writes
-    ///
-    /// If `used` is larger than the given grant, this function will panic.
-    #[inline(always)]
-    pub fn release(&mut self, used: usize, grant: GrantR) {
-        unsafe { self.bbq.as_mut().release(used, grant) }
-    }
-
-    /// Returns the size of the backing storage.
-    ///
-    /// This is the maximum number of bytes that can be stored in this queue.
-    pub fn capacity(&self) -> usize {
-        unsafe { self.bbq.as_ref().capacity() }
-    }
-}
-
-/// Statically allocate a `BBQueue` singleton object with a given size (in bytes).
-///
-/// This function must only ever be called once in the same place
-/// per function. For example:
-///
-/// ```rust,ignore
-/// // this creates two separate/distinct buffers, and is an acceptable usage
-/// let bbq1 = bbq!(1024).unwrap(); // Returns Some(BBQueue)
-/// let bbq2 = bbq!(1024).unwrap(); // Returns Some(BBQueue)
-///
-/// // this is a bad example, as the same instance is executed twice!
-/// // On the first call, this macro will return `Some(BBQueue)`. On the
-/// // second call, this macro will return `None`!
-/// for _ in 0..2 {
-///     let _ = bbq!(1024).unwrap();
-/// }
-/// ```
-///
-/// If you are using this in a cortex-m microcontroller system,
-/// consider using the cortex_m_bbq!() macro instead.
-#[macro_export]
-macro_rules! bbq {
-    ($expr:expr) => {{
-        use core::sync::atomic::{AtomicBool, Ordering};
-
-        static TAKEN: AtomicBool = AtomicBool::new(false);
-
-        if TAKEN.compare_and_swap(false, true, Ordering::AcqRel) {
-            None
-        } else {
-            unsafe { $crate::unchecked_bbq!($expr) }
-        }
-    }};
-}
-
-/// Like the `bbq!()` macro, but wraps the initialization in a cortex-m
-/// "critical section" by disabling interrupts
-#[cfg_attr(feature = "cortex-m", macro_export)]
-#[allow(unused_macros)]
-macro_rules! cortex_m_bbq {
-    ($expr:expr) => {
-        cortex_m::interrupt::free(|_| unsafe { $crate::unchecked_bbq!($expr) })
-    };
-}
-
-/// This macro does try to provide similar guarantees as `bbq!()`,
-/// but is not thread safe.
-#[macro_export]
-macro_rules! unchecked_bbq {
-    ($expr:expr) => {{
-        static mut BUFFER: [u8; $expr] = [0u8; $expr];
-
-        // Really, this shouldn't be an option. But for now, there is
-        // no stable way to get an uninitialized static, so we pay the
-        // option cost for compatibility
-        static mut BBQ: Option<BBQueue> = None;
-
-        if BBQ.is_some() {
-            None
-        } else {
-            BBQ = Some(BBQueue::unpinned_new(&mut BUFFER));
-            BBQ.as_mut()
-        }
-    }};
-}
+//         if BBQ.is_some() {
+//             None
+//         } else {
+//             BBQ = Some(BBQueue::unpinned_new(&mut BUFFER));
+//             BBQ.as_mut()
+//         }
+//     }};
+// }
